IntegratedServer/bot/bot_main.py
# IntegratedServer/bot/bot_main.py
# Telegram Bot Main Logic

import logging
from typing import Any
from config import BOT_TOKEN, BOT_USERNAME
from telegram import Update
from telegram.ext import Application, ContextTypes

logger = logging.getLogger(__name__)

async def setup_bot():
    """Setup and start Telegram bot"""
    
    if not BOT_TOKEN:
        logger.error("❌ BOT_TOKEN not configured!")
        return None
    
    logger.info(f"🤖 Initializing Telegram Bot: @{BOT_USERNAME}")
    
    # Create application
    app = Application.builder().token(BOT_TOKEN).build()
    
    logger.info("✅ Bot handlers will be registered here")
    # TODO: Register handlers from original bot code
    # - start_handlers
    # - conversation_handlers
    # - message_handlers
    # - etc.
    
    # Start bot
    logger.info("🚀 Bot is ready and listening...")
    
    await app.run_polling()

def register_handlers(app: Application[Any, Any, Any, Any, Any, Any]):
    """Register all bot handlers"""
    logger.info("📝 Registering bot handlers...")
    
    from telegram.ext import CommandHandler
    
    async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        await update.message.reply_text("Hello! MovieZone Bot is running!")
    
    app.add_handler(CommandHandler("start", start))
    
    logger.info("✅ Handlers registered")
