IntegratedServer/requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Telegram Bot
python-telegram-bot==20.3
APScheduler==3.10.4

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.1

# ORM & Database Tools
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & API
requests==2.31.0

# Utilities
python-dateutil==2.8.2
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.1

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Production
Werkzeug==3.0.1
